+++ src/app/user-dasboard/user-profile/user-profile.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { collection, collectionData, deleteDoc, doc, Firestore, query, updateDoc, where } from '@angular/fire/firestore';
 import { ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-profile',
@@ -8,6 +9,7 @@ import { ActivatedRoute } from '@angular/router';
   styleUrls: ['./user-profile.component.css']
 })
 export class UserProfileComponent {
+  loader : boolean = false;
   userDetails: any;
   myRecipes: any;
   userId: any = JSON.parse(localStorage.getItem("userDetails") as string)
@@ -21,8 +23,9 @@ export class UserProfileComponent {
   getLoggedinUser() {
     let collectionName = collection(this.firestore, "users")
     let firebaseQuery = query(collectionName, where("userId", "==", this.userId))
-
+    this.loader = true
     collectionData(firebaseQuery, { idField: "id" }).subscribe((user) => {
+      this.loader = false;
       console.log(user)
       this.userDetails = user[0]
     })
@@ -32,7 +35,7 @@ updateLoggedinUser(){
   let data = { name: this.userDetails.name, 
                phoneNumber: this.userDetails.phoneNumber};
   updateDoc(doc(this.firestore, "users" + "/" + this.userDetails.id), data).then(res => {
-    console.log("User profile updated successfully!");
+    Swal.fire("User profile updated successfully!");
   }).catch((error) => {
     console.error("Error updating user profile:", error);
   });
@@ -40,21 +43,41 @@ updateLoggedinUser(){
 
   getMyRecipes(){
     const firebaseCollectionName = collection(this.firestore, "recipe")
+    this.loader = true
     const firebaseQurey = query(firebaseCollectionName,where("userId","==", JSON.parse(localStorage.getItem("userDetails")as string)))
     collectionData(firebaseQurey, { idField: 'id'}).subscribe((myRecipes: any) => {
+      this.loader = false
       console.log("Fetched Recipe", myRecipes);
       this.myRecipes = myRecipes;
     })
   }
 
-  delete(id:any) {
-    deleteDoc(doc(this.firestore, "recipe/" + id)).then(res => {
-      alert("Your recipe has been deleted")
-      console.log(res)
-    }).catch(err =>{
-      console.log(err)
-    })
-  }
+  delete(id:any){
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      console.log(result)
+      if (result.isConfirmed) {
+        deleteDoc(doc(this.firestore, "recipe/" + id)).then(res => {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+          console.log(res)
+        }).catch(err =>{
+          console.log(err)
+        })
+      }
+    });
+    
+   }
 
 
 }
